feat(product): Knowledge Base / Sessions tabs + one shared composer
- Product splits into two tabs: Knowledge Base (the page tree + content)
  and Sessions (chat history with the digital product owner).
- Sessions is the Claude Desktop shape: session list left, conversation
  right, composer below. The composer is THE component — extracted from
  the floating bar (ask-composer.tsx) and reused verbatim: same
  paste-to-intake chips, same shielded input, same /ask wiring, generic
  'chat with the digital product owner' placeholder (no ask/follow-up
  assumption). The floating bar steps aside on the Sessions route (the
  tam-os duplicate-input rule) and hosts the same composer everywhere
  else.
- Knowledge Base tree unified on the semi-structured design: charter and
  active leave the legacy pinned block and join a 'context' section with
  CHARTER/ACTIVE reference ids, consistent with every other section; the
  conversations section moves out of the tree (Sessions owns it).
- thread-view superseded by ask/thread-messages (display-only, shared by
  the floating panel's sibling and the Sessions workspace).
